+++ test/bootstrap.el
@@ -36,10 +36,12 @@
 
   (mapc (lambda (pkg)
           (unless (package-installed-p pkg)
-            (package-install pkg)))
+            (package-refresh-contents) (package-install pkg)))
         pkgs)
 
   (add-hook 'kill-emacs-hook
-            `(lambda () (delete-directory ,user-emacs-directory t))))
+            `(lambda ()
+               (unless (boundp 'emacs-lsp-ci)
+                 (delete-directory ,user-emacs-directory t)))))
 
 ;;; bootstrap.el ends here