+++ CHANGELOG.md
@@ -5,7 +5,26 @@ All notable changes to this project will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 
+### 2021-03-18
+
+* Add support for JSON.
+
+### 2021-03-17
+
+* Add support for HTML.
+* Add support for XML.
+
+### 2021-02-05
+
+* Fixed `issue from c-style` overlap node.
+
 ### 2021-02-04
+
+* Expand `c-style` parser to the beginning of line.
+* Add support for multiline comment in Lua.
+* Add `else` region folding.
+
+### 2021-02-03
 
 * Fixed filtering error if previous line is comment/string block.
 * Add Ruby support.